refactored out _parse_snapshot for _file_client.py and async
