+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/_file_client.py
@@ -25,7 +25,6 @@ from ._file_client_helpers import (
     _format_url,
     _from_file_url,
     _get_ranges_options,
-    _parse_snapshot,
     _parse_url,
     _upload_range_from_url_options
 )
@@ -33,7 +32,7 @@ from ._generated import AzureFileStorage
 from ._generated.models import FileHTTPHeaders
 from ._lease import ShareLeaseClient
 from ._models import HandlesPaged
-from ._parser import _get_file_permission, _datetime_to_str
+from ._parser import _datetime_to_str, _get_file_permission, _parse_snapshot
 from ._serialize import (
     get_access_conditions,
     get_api_version,
